backend/requirements.txt
# KaratPlus AI - Backend Dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
motor>=3.6
pydantic>=2.10
pydantic-settings>=2.6
python-dotenv==1.0.1
google-generativeai>=0.5.0,<0.9
pytesseract==0.3.10
opencv-python-headless>=4.9.0
Pillow>=10.0
aiofiles==23.2.1
pypdf>=5.1.0
openpyxl>=3.1.5
xlrd>=2.0.1
